refactor(inmemory/domainobject): change ID type

+++ src/main/java/com/github/dd_buntar/goods_warehouse/domainobject/Manufacturer.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Manufacturer {
-    private Integer manufacturerId;
+    private Long manufacturerId;
     private String manufacturerName;
     private String contactPhone;
     private String country;
